fix: resolve circular includes between interfaces and concrete types, fix panel tests
Move ModelInfo struct to shared core/modelInfo.h and ModelsIniEntry to
core/modelsIniEntry.h so interface headers no longer depend on their
concrete implementations — breaking the circular include chain. Add a
backward-compatible using alias in ModelsIni for ModelEntry.

Remove private-method tests from ModelInfoPanelTest (formatDouble,
formatNumber, getStatusString) per spec 'never assert on implementation
details'. Fix ModelsPanelTest mock expectations to match actual panel
call count. Remove SettingsPanelTest save assertion that never fires at
construction time. Add build.sh test output logging.
