+++ tests/mocks/MockModelsIni.h
@@ -7,7 +7,7 @@
 class MockModelsIni : public IModelsIni
 {
   public:
-    MOCK_CONST_METHOD0(getUniqueModelEntries, std::vector<ModelsIni::ModelEntry>());
+    MOCK_CONST_METHOD0(getUniqueModelEntries, std::vector<ModelsIniEntry>());
     MOCK_CONST_METHOD1(getModelPath, std::string(const std::string &sectionName));
     MOCK_CONST_METHOD1(getPresetsForModel, std::vector<Config::ModelPreset>(const std::string &modelPath));
     MOCK_METHOD1(savePreset, bool(const Config::ModelPreset &preset));
