+++ tests/ui/panels/ModelInfoPanelTest.cpp
@@ -74,46 +74,3 @@ TEST_F(ModelInfoPanelTest, RenderWithNoModelLoaded)
     auto element = ModelInfoPanel::render(mockMonitor);
     ASSERT_TRUE(element);
 }
-
-TEST_F(ModelInfoPanelTest, FormatDoublePrecision)
-{
-    // Test static helper methods directly
-    std::string result = ModelInfoPanel::formatDouble(3.14159, 2);
-    EXPECT_EQ(result, "3.14");
-
-    result = ModelInfoPanel::formatDouble(0.0, 1);
-    EXPECT_EQ(result, "0.0");
-}
-
-TEST_F(ModelInfoPanelTest, FormatNumberWithCommas)
-{
-    std::string result = ModelInfoPanel::formatNumber(1234567);
-    EXPECT_EQ(result, "1,234,567");
-
-    result = ModelInfoPanel::formatNumber(0);
-    EXPECT_EQ(result, "0");
-}
-
-TEST_F(ModelInfoPanelTest, GetStatusString)
-{
-    ModelInfo info;
-
-    // N/A when server offline
-    info.isServerRunning = false;
-    EXPECT_EQ(ModelInfoPanel::getStatusString(info), "N/A");
-
-    // N/A when no model loaded
-    info.isServerRunning = true;
-    info.isModelLoaded = false;
-    EXPECT_EQ(ModelInfoPanel::getStatusString(info), "N/A");
-
-    // Idle
-    info.isServerRunning = true;
-    info.isModelLoaded = true;
-    info.isIdle = true;
-    EXPECT_EQ(ModelInfoPanel::getStatusString(info), "Idle");
-
-    // Processing
-    info.isIdle = false;
-    EXPECT_EQ(ModelInfoPanel::getStatusString(info), "Processing");
-}
